Fixed size_t compiler warnings in the sampleRun() function.

+++ Name Generator/Source.cpp
@@ -251,8 +251,8 @@ void sampleRun() {
 	//read in names and store in vector until the end of the file
 	while (fin >> nextName) {
 		//keep track of length of longest word
-		if (max_length < nextName.length())
-			max_length = nextName.length();
+		if (max_length < (int)nextName.length())
+			max_length = (int)nextName.length();
 		//ensure word has at least one vowel
 		if (hasVowel(nextName))
 			names.push_back(nextName);
